add: `semantic` workflow versioned approach
