+++ warehouse/oso_agent/oso_agent/workflows/text2sql/semantic_iters/__init__.py
@@ -1,0 +1,7 @@
+from .v18_08_25.i0 import SemanticText2SQLWorkflow as SemanticText2SQLWorkflowV0
+from .v27_08_25.i1 import SemanticText2SQLWorkflow as SemanticText2SQLWorkflowV1
+
+__all__ = [
+    "SemanticText2SQLWorkflowV0",
+    "SemanticText2SQLWorkflowV1",
+]
