+++ warehouse/oso_agent/oso_agent/workflows/text2sql/semantic_iters/v18_08_25/README.md
@@ -1,0 +1,4 @@
+# iteration 0
+
+This is the base workflow. Performance will be measured relative to this
+workflow.
